Implemented method of combining jewels.
Added variable to indicate if jewel is to be rendered or not in Jewel
class. Also added its function modifier and access.

Added two functions to Grid class. The first is to find combinations in
neighbouring jewels to one indicated. The second is to set the jewels
to be combined to disappear.

Modified event handler to make use of functions created in Grid class.